Improve data types for GPS navigation messages

+++ src/core/libs/gnss_sdr_supl_client.cc
@@ -220,12 +220,12 @@ void gnss_sdr_supl_client::read_supl_data()
         {
             gps_utc.d_A0 = static_cast<double>(assist.utc.a0) * pow(2.0, -30);
             gps_utc.d_A1 = static_cast<double>(assist.utc.a1) * pow(2.0, -50);
-            gps_utc.d_DeltaT_LS = static_cast<double>(assist.utc.delta_tls);
-            gps_utc.d_DeltaT_LSF = static_cast<double>(assist.utc.delta_tlsf);
-            gps_utc.d_t_OT = static_cast<double>(assist.utc.tot) * pow(2.0, 12);
-            gps_utc.i_DN = static_cast<double>(assist.utc.dn);
-            gps_utc.i_WN_T = static_cast<double>(assist.utc.wnt);
-            gps_utc.i_WN_LSF = static_cast<double>(assist.utc.wnlsf);
+            gps_utc.d_DeltaT_LS = static_cast<int32_t>(assist.utc.delta_tls);
+            gps_utc.d_DeltaT_LSF = static_cast<int32_t>(assist.utc.delta_tlsf);
+            gps_utc.d_t_OT = static_cast<int32_t>(assist.utc.tot) * pow(2.0, 12);
+            gps_utc.i_DN = static_cast<int32_t>(assist.utc.dn);
+            gps_utc.i_WN_T = static_cast<int32_t>(assist.utc.wnt);
+            gps_utc.i_WN_LSF = static_cast<int32_t>(assist.utc.wnlsf);
             gps_utc.valid = true;
         }
 
@@ -268,7 +268,7 @@ void gnss_sdr_supl_client::read_supl_data()
                     gps_almanac_iterator->second.d_OMEGA0 = static_cast<double>(a->OMEGA_0) * pow(2.0, -23);
                     gps_almanac_iterator->second.d_sqrt_A = static_cast<double>(a->A_sqrt) * pow(2.0, -11);
                     gps_almanac_iterator->second.d_OMEGA_DOT = static_cast<double>(a->OMEGA_dot) * pow(2.0, -38);
-                    gps_almanac_iterator->second.i_Toa = static_cast<int32_t>(a->toa) * pow(2.0, 12);
+                    gps_almanac_iterator->second.i_Toa = static_cast<int32_t>(a->toa * pow(2.0, 12));
                     gps_almanac_iterator->second.d_e_eccentricity = static_cast<double>(a->e) * pow(2.0, -21);
                     gps_almanac_iterator->second.d_M_0 = static_cast<double>(a->M0) * pow(2.0, -23);
                 }
